Fix Command Topics and Verbs
